+++ src/backend/controllers/QuotationsController.js
@@ -1,5 +1,6 @@
 import express from "express";
 import * as QuotationsService from "../services/QuotationsService.js";
+import { NotFoundErrorName } from "../errors/NotFoundError.js";
 
 const router = express.Router();
